Add 'sfdiskM' command.
This command is a saner interface to partitioning.

All partition sizes are specified in megabytes (not cylinders).
You don't need to specify the cyls/heads/secs parameters.

All the test code has been updated to use this, so it is now
CHS-independent (eg. when CHS changes as between IDE and virtio).