tests: Fix read_file test.
The test ignored the fact that this function returns a (char *, size_t)
pair, and just treated the char * as a string.  This of course would
fail if the string didn't happen to be nul-terminated.

The tests are updated to add a new TestOutputBuffer type which should
be used with functions that return RBufferOut.