Sys::Guestfs::Lib::open_guest: Remove freeform parameters.
Remove the ability to pass freeform parameters to Sys::Virt->new.
We don't use it, it makes the code more complex to modify, and
indeed there are no other args that Sys::Virt->new supports so
this would never be used.

Also change $readwrite to $rw to match parameter name.